requirements.txt
amqp==5.0.6
antiorm==1.2.1
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==21.1.0
arrayfire==3.6.20181017
arrow==1.2.1
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
automium==0.2.6
automium-web==0.1.1
awsebcli==3.20.2
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
binaryornot==0.4.4
boto3==1.20.8
botocore==1.23.8
Brotli==1.0.9
cached-property==1.5.2
celery==5.2.1
cement==3.0.4
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
cocos==0.2.2
cocos2d==0.6.9
colorama==0.4.4
config==0.5.1
ConfigArgParse==1.5.3
constantly==15.1.0
contexttimer==0.3.3
cookiecutter==1.7.3
coverage==6.1.2
crispy-tailwind==0.5.0
cryptography==35.0.0
daphne==3.0.2
db==0.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.3
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.9
django-admin-interface==0.18.2
django-allauth==0.46.0
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-colorfield==0.4.5
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-crum==0.7.9
django-dynamic-fixture==3.1.2
django-environ==0.8.1
django-factory-boy==1.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-pytest==0.2.0
django-redis==5.0.0
django-rest-auth==0.9.5
django-storages==1.12.3
django-taggit==2.0.0
django-tailwind==2.2.2
django-timezone-field==4.2.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.8.2
filelock==3.4.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Factory==0.1.0.dev0
future==0.18.2
gevent==21.8.0
geventhttpclient==1.5.3
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
iniconfig==1.1.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
kombu==5.2.2
legacy==0.1.6
locust==2.5.0
loky==3.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
mpmath==1.2.1
msgpack==1.0.2
multiprocess==0.70.12.2
mutagen==1.45.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.0.3
numpy==1.21.4
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pathos==0.2.8
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pox==0.3.0
poyo==0.5.0
ppft==1.6.6.4
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
psycopg2==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyglet==1.5.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyPDF2==1.26.0
pypiwin32==223
pytest==6.2.5
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
pywin32==302
PyYAML==6.0
pyzmq==22.3.0
redis==4.0.1
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
reverse==0.1.0
roundrobin==0.0.2
s3transfer==0.5.0
scipy==1.7.2
selenium==4.0.0
semantic-version==2.8.5
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
state==0.1.1.dev0
stripe==2.63.0
sympy==1.9
termcolor==1.1.0
tests==0.7
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
wsproto==1.0.0
xhtml2pdf==0.2.5
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
